chore: Extracted Map component from MapPage
