+++ src/components/Toolbar/index.tsx
@@ -1,0 +1,5 @@
+const Toolbar = () => {
+  return <div>Toolbar</div>;
+};
+
+export default Toolbar;
